+++ src/infrastructure/redis/domain/decks/deck.cache-repository.ts
@@ -1,0 +1,16 @@
+import { DeckReadModel } from '@infrastructure/mongo/domain/decks/deck.read-model';
+import { BaseRedisRepository } from '@infrastructure/redis/base-redis-repository';
+import { RedisClient } from 'redis';
+
+interface Dependencies {
+  redisClient: RedisClient;
+}
+
+export class DeckCacheRepository extends BaseRedisRepository<DeckReadModel[]> {
+  constructor(dependencies: Dependencies) {
+    super({
+      redisClient: dependencies.redisClient,
+      persistedKeyPrefix: 'DECKS_',
+    });
+  }
+}