+++ src/infrastructure/redis/redis-client.ts
@@ -1,0 +1,8 @@
+import { createClient } from 'redis';
+
+require('dotenv').config();
+
+export const createRedisClient = () =>
+  createClient({
+    port: Number(process.env.REDIS_PORT),
+  });